+++ bundle/src/main/java/com/adobe/acs/commons/mcp/impl/processes/renovator/Util.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.adobe.acs.commons.mcp.impl.processes.reorganizer;
+package com.adobe.acs.commons.mcp.impl.processes.renovator;
 
 import com.adobe.acs.commons.fam.actions.Actions;
 import com.day.cq.replication.ReplicationStatus;
